backend/requirements.txt
flake8==3.5.0
Flask==1.0.2
Flask-Cors==3.0.4
psycopg2-binary==2.7.5
pytest==3.5.1
pytest-cov==2.5.1
python-dotenv==0.9.0
requests==2.18.4
SQLAlchemy==1.2.7
